pyproject.toml
[project]
name = "travel-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.8.0",
    "aioredis>=2.0.0",
    "alembic>=1.12.0",
    "anthropic>=0.30.0",
    "asyncpg>=0.28.0",
    "beautifulsoup4>=4.13.4",
    "black>=23.0.0",
    "duckduckgo-search>=8.1.1",
    "fastapi>=0.100.0",
    "httpx>=0.25.0",
    "isort>=5.0.0",
    "jsonschema>=4.0.0",
    "langchain>=0.2.0",
    "langchain-community>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-groq>=0.3.2",
    "langchain-tavily>=0.2.11",
    "langgraph>=0.2.0",
    "langgraph-supervisor>=0.0.28",
    "openai>=1.40.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dateutil>=2.8.0",
    "python-dotenv>=1.0.0",
    "redis>=5.0.0",
    "sqlalchemy>=2.0.0",
    "structlog>=23.0.0",
    "together>=1.0.0",
    "uvicorn[standard]>=0.20.0",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.13",
]
